+++ payment-service/src/main/java/com/hedvig/paymentservice/domain/swish/commands/SwishPayoutTransactionFailedCommand.kt
@@ -6,5 +6,8 @@ import java.util.UUID
 class SwishPayoutTransactionFailedCommand(
     @TargetAggregateIdentifier
     val transactionId: UUID,
-    val memberId: String
+    val memberId: String,
+    val errorCode :String?,
+    val errorMessage :String?,
+    val additionalInformation :String?
 )
