+++ payment-service/src/main/java/com/hedvig/paymentservice/services/swish/StartPayoutResponse.kt
@@ -1,0 +1,9 @@
+package com.hedvig.paymentservice.services.swish
+
+sealed class StartPayoutResponse {
+    object Success: StartPayoutResponse()
+    data class Failed(
+        val exceptionMessage: String?,
+        val status: Int?
+    ) : StartPayoutResponse()
+}
