+++ payment-service/src/main/java/com/hedvig/paymentservice/services/swish/SwishService.kt
@@ -1,57 +1,14 @@
 package com.hedvig.paymentservice.services.swish
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import feign.Client
+import com.hedvig.paymentservice.services.swish.client.SwishClient
 import feign.FeignException
-import org.apache.http.conn.ssl.DefaultHostnameVerifier
-import org.apache.http.ssl.SSLContexts
-import org.bouncycastle.asn1.pkcs.PrivateKeyInfo
-import org.bouncycastle.openssl.PEMParser
-import org.bouncycastle.openssl.jcajce.JcaPEMKeyConverter
-import org.javamoney.moneta.Money
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.cloud.openfeign.FeignClient
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.http.ResponseEntity
 import org.springframework.stereotype.Service
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
-import java.io.File
-import java.io.FileReader
-import java.math.BigDecimal
-import java.security.MessageDigest
-import java.security.PrivateKey
-import java.security.Signature
 import java.time.LocalDateTime
-import java.util.Base64
 import java.util.UUID
 import javax.money.MonetaryAmount
-import javax.net.ssl.SSLSocketFactory
-
-@Configuration
-@ConfigurationProperties(prefix = "swish")
-class SwishConfigurationProperties {
-    lateinit var tlsCertPath: String
-    lateinit var tlsCertPassword: String
-    lateinit var signingPrivatePemPath: String
-    lateinit var signingCertificateSerialNumber: String
-    lateinit var callbackUrl: String
-}
-
-sealed class StartPayoutResponse {
-    object Success: StartPayoutResponse()
-    data class Failed(
-        val exceptionMessage: String?,
-        val status: Int?
-    ) : StartPayoutResponse()
-}
 
 @Service
 class SwishService(
@@ -61,6 +18,7 @@ class SwishService(
 ) {
     fun startPayout(
         transactionId: UUID,
+        memberId: String,
         payeeAlias: String,
         payeeSSN: String,
         amount: MonetaryAmount,
@@ -68,17 +26,17 @@ class SwishService(
         instructionDate: LocalDateTime,
     ): StartPayoutResponse {
         val payload = PayoutPayload(
-            transactionId.toString().replace("-", "").toUpperCase(),
-            transactionId,
-            payeeAlias,
-            payeeSSN,
-            String.format("%.2f", amount),
-            message,
-            instructionDate.toString(),
-            properties.signingCertificateSerialNumber
+            payoutInstructionUUID = transactionId.toString().replace("-", "").toUpperCase(),
+            payerPaymentReference = memberId,
+            payeeAlias = payeeAlias,
+            payeeSSN = payeeSSN,
+            amount = String.format("%.2f", amount),
+            message = message,
+            instructionDate = instructionDate.toString(),
+            signingCertificateSerialNumber = properties.signingCertificateSerialNumber
         )
         val json = objectMapper.writeValueAsString(payload)
-        val signature = SwishSignature.createSignature(json, properties.signingPrivatePemPath)
+        val signature = SwishSignatureCreator.createSignature(json, properties.signingPrivatePemPath)
         val req = PayoutRequest(payload, signature, properties.callbackUrl)
         return try {
             client.payout(req)
@@ -96,7 +54,7 @@ class SwishService(
 
     data class PayoutPayload(
         val payoutInstructionUUID: String,
-        val payerPaymentReference: UUID,
+        val payerPaymentReference: String,
         val payeeAlias: String,
         val payeeSSN: String,
         val amount: String,
@@ -107,115 +65,5 @@ class SwishService(
         val payerAlias: String = "1235261086"
         val currency: String = "SEK"
         val payoutType: String = "PAYOUT"
-    }
-
-    private val log: Logger = LoggerFactory.getLogger(this::class.java)
-
-}
-
-object SwishSignature {
-
-    fun createSignature(payload: String, signingPrivatePemPath: String): String {
-        val msgBytes = payload.toByteArray()
-        val md = MessageDigest.getInstance("SHA-512")
-        val hashValue = md.digest(msgBytes);
-
-        val sig = Signature.getInstance("NONEwithRSA")
-        val privateKey = loadPrivateKey(signingPrivatePemPath)
-        sig.initSign(privateKey)
-        sig.update(hashValue)
-        val signatureBytes = sig.sign()
-        return Base64.getEncoder().encodeToString(signatureBytes)
-    }
-
-    fun loadPrivateKey(signingPrivatePemPath: String): PrivateKey {
-        val pemParser = PEMParser(FileReader(File(signingPrivatePemPath)))
-        val privateKeyInfoAny = pemParser.readObject()
-        val converter = JcaPEMKeyConverter()
-
-        return converter.getPrivateKey(privateKeyInfoAny as PrivateKeyInfo)
-    }
-
-}
-
-@RestController
-@RequestMapping(path = ["/swish/"])
-class SwishController(
-    val swishService: SwishService
-) {
-
-    @GetMapping("payout")
-    fun payout() {
-        swishService.startPayout(
-            transactionId = UUID.randomUUID(),
-            payeeAlias = "46726738711",
-            payeeSSN = "198607209882",
-            amount = Money.of(10, "SEK"),
-            message = "message",
-            instructionDate = LocalDateTime.now(),
-        )
-    }
-
-    @PostMapping("callback")
-    fun callback(callback: Callback) {
-        log.info(callback.toString())
-    }
-
-    private val log: Logger = LoggerFactory.getLogger(this::class.java)
-
-    data class Callback(
-        val payoutInstructionUUID: String,
-        val payerPaymentReference: String,
-        val callbackUrl: String,
-        val payerAlias: String,
-        val payeeAlias: String,
-        val payeeSSN: String,
-        val amount: String,
-        val currency: String,
-        val message: String,
-        val status: String,
-        val payoutType: String,
-        val dateCreated: LocalDateTime,
-        val datePaid: LocalDateTime,
-        val errorCode: String?,
-        val errorMessage: String?,
-        val additionalInformation: String?
-    )
-}
-
-@FeignClient(
-    name = "swishClient",
-    url = "\${hedvig.external.swish.baseurl:https://staging.getswish.pub.tds.tieto.com/cpc-swish/}",
-    configuration = [SwishFeignConfiguration::class]
-)
-interface SwishClient {
-    @PostMapping("/api/v1/payouts/")
-    fun payout(
-        @RequestBody payloadRequest: SwishService.PayoutRequest
-    ): ResponseEntity<*>
-}
-
-class SwishFeignConfiguration {
-
-    @Autowired
-    lateinit var properties : SwishConfigurationProperties
-
-    private val log: Logger = LoggerFactory.getLogger(this::class.java)
-
-    @Bean
-    fun feignClient(): Client {
-        val sslSocketFactory = getSSLSocketFactory()
-        return Client.Default(sslSocketFactory, DefaultHostnameVerifier())
-    }
-
-    private fun getSSLSocketFactory(): SSLSocketFactory? = try {
-        val sslContext = SSLContexts
-            .custom()
-            .loadKeyMaterial(File(properties.tlsCertPath), properties.tlsCertPassword.toCharArray(), properties.tlsCertPassword.toCharArray())
-            .build()
-        sslContext.socketFactory
-    } catch (exception: Exception) {
-        log.error("Failed to set up swish TLS")
-        null
     }
 }
